Add reboot message after /uninstall and use MOVEFILE_DELAY_UNTIL_REBOOT
I'm giving up on uninstalling without a system reboot.
Windows requires me to stop the driver service before it unloads the driver from memory, but stopping is not possible for a filter driver.
You are required to restart the entire stack where the filter driver is attached to, but this is neither possible for legacy devices like COM ports.

On top of that, Windows 10 locks the driver file in memory and doesn't allow to delete it.
MoveFileExW with MOVEFILE_DELAY_UNTIL_REBOOT is the only option here.